+++ FlatFiles.Test/TypeDeductionTester.cs
@@ -14,21 +14,22 @@ namespace FlatFiles.Test
         [TestMethod]
         public void ShouldDeduceSchemaForType()
         {
-            const string data = @"Id,Name,CreatedOn,IsActive,VisitCount
-1,Bob,2018-07-01,true,1
-2,John,2018-07-2,false,
-3,Susan,2018-07-03,false,10
-";
-            var stringReader = new StringReader(data);
-            var reader = SeparatedValueTypeMapper.GetAutoMappedReader<Person>(stringReader);
-            var results = reader.ReadAll().ToArray();
-            Assert.AreEqual(3, results.Length, "The wrong number of records were read.");
+            var stringWriter = new StringWriter();
+            var writer = SeparatedValueTypeMapper.GetAutoMappedWriter<Person>(stringWriter);
             var expected = new[]
             {
                 new Person() { Id = 1, Name = "Bob", CreatedOn = new DateTime(2018, 07, 01), IsActive = true, VisitCount = 1 },
                 new Person() { Id = 2, Name = "John", CreatedOn = new DateTime(2018, 07, 02), IsActive = false, VisitCount = null },
                 new Person() { Id = 3, Name = "Susan", CreatedOn = new DateTime(2018, 07, 03), IsActive = false, VisitCount = 10 }
             };
+            writer.WriteAll(expected);
+            string output = stringWriter.ToString();
+
+            var stringReader = new StringReader(output);
+            var reader = SeparatedValueTypeMapper.GetAutoMappedReader<Person>(stringReader);
+            var results = reader.ReadAll().ToArray();
+
+            Assert.AreEqual(3, results.Length, "The wrong number of records were read.");
             AssertEqual(expected, results, 0);
             AssertEqual(expected, results, 1);
             AssertEqual(expected, results, 2);
@@ -37,25 +38,26 @@ namespace FlatFiles.Test
         [TestMethod]
         public async Task ShouldDeduceSchemaForTypeAsync()
         {
-            const string data = @"Id,Name,CreatedOn,IsActive,VisitCount
-1,Bob,2018-07-01,true,1
-2,John,2018-07-2,false,
-3,Susan,2018-07-03,false,10
-";
-            var stringReader = new StringReader(data);
-            var reader = await SeparatedValueTypeMapper.GetAutoMappedReaderAsync<Person>(stringReader);
-            var results = new List<Person>();
-            while (await reader.ReadAsync())
-            {
-                results.Add(reader.Current);
-            }
-            Assert.AreEqual(3, results.Count, "The wrong number of records were read.");
+            var stringWriter = new StringWriter();
+            var writer = SeparatedValueTypeMapper.GetAutoMappedWriter<Person>(stringWriter);
             var expected = new[]
             {
                 new Person() { Id = 1, Name = "Bob", CreatedOn = new DateTime(2018, 07, 01), IsActive = true, VisitCount = 1 },
                 new Person() { Id = 2, Name = "John", CreatedOn = new DateTime(2018, 07, 02), IsActive = false, VisitCount = null },
                 new Person() { Id = 3, Name = "Susan", CreatedOn = new DateTime(2018, 07, 03), IsActive = false, VisitCount = 10 }
             };
+            await writer.WriteAllAsync(expected);
+            string output = stringWriter.ToString();
+
+            var stringReader = new StringReader(output);
+            var reader = await SeparatedValueTypeMapper.GetAutoMappedReaderAsync<Person>(stringReader);
+            var results = new List<Person>();
+            while (await reader.ReadAsync())
+            {
+                results.Add(reader.Current);
+            }
+
+            Assert.AreEqual(3, results.Count, "The wrong number of records were read.");
             AssertEqual(expected, results, 0);
             AssertEqual(expected, results, 1);
             AssertEqual(expected, results, 2);